feat(raw-material): #829 group edit as a route drawer + ActionMenu + material-type picker
- Group detail "Edit" moves from an inline FocusModal to an ActionMenu (…) that
  opens a RouteDrawer at /raw-material-groups/:id/edit (EditGroupForm).
- Add a material-type CATEGORY picker (CategorySearch) to the group edit form.
  The update endpoint now accepts a category name (material_type) and find-or-
  creates it → material_type_id, mirroring the raw-material create flow.
