+++ apps/backend/src/admin/components/raw-material-groups/group-sections.tsx
@@ -11,7 +11,6 @@ import {
   Select,
   Table,
   Text,
-  Textarea,
   FocusModal,
   createDataTableColumnHelper,
   toast,
@@ -29,7 +28,6 @@ import {
   useLinkGroupColors,
   useCreateGroupOrder,
   useRawMaterialGroupOrders,
-  useUpdateRawMaterialGroup,
 } from "../../hooks/api/raw-material-groups"
 import { useInventoryWithRawMaterials } from "../../hooks/api/raw-materials"
 import { useStockLocations } from "../../hooks/api/stock_location"
@@ -74,7 +72,19 @@ export const GroupGeneralSection = ({ group }: { group: RawMaterialGroup }) => {
           >
             {group.status || "Active"}
           </Badge>
-          <EditGroupModal group={group} locations={stock_locations} />
+          <ActionMenu
+            groups={[
+              {
+                actions: [
+                  {
+                    icon: <PencilSquare />,
+                    label: "Edit",
+                    to: `/raw-material-groups/${group.id}/edit`,
+                  },
+                ],
+              },
+            ]}
+          />
         </div>
       </div>
       {rows.map(([label, value]) => (
@@ -89,224 +99,6 @@ export const GroupGeneralSection = ({ group }: { group: RawMaterialGroup }) => {
         </div>
       ))}
     </Container>
-  )
-}
-
-// ---------------------------------------------------------------------------
-// Edit group globals (#829) — set once, inherited fill-blank by new colors.
-// ---------------------------------------------------------------------------
-
-const UNIT_OPTIONS = ["Meter", "Yard", "Kilogram", "Gram", "Piece", "Roll", "Other"]
-const STATUS_OPTIONS = ["Active", "Discontinued", "Under_Review", "Development"]
-
-const EditGroupModal = ({
-  group,
-  locations,
-}: {
-  group: RawMaterialGroup
-  locations: any[]
-}) => {
-  const [open, setOpen] = useState(false)
-  const { mutateAsync, isPending } = useUpdateRawMaterialGroup(group.id)
-
-  const [name, setName] = useState(group.name || "")
-  const [status, setStatus] = useState(group.status || "Active")
-  const [composition, setComposition] = useState(group.composition || "")
-  const [unit, setUnit] = useState(group.unit_of_measure || "Other")
-  const [locationId, setLocationId] = useState(group.stock_location_id || "")
-  const [unitCost, setUnitCost] = useState(
-    group.unit_cost != null ? String(group.unit_cost) : ""
-  )
-  const [currency, setCurrency] = useState(group.cost_currency || "")
-  const [leadTime, setLeadTime] = useState(
-    group.lead_time_days != null ? String(group.lead_time_days) : ""
-  )
-  const [moq, setMoq] = useState(
-    group.minimum_order_quantity != null
-      ? String(group.minimum_order_quantity)
-      : ""
-  )
-  const [specs, setSpecs] = useState(
-    group.specifications ? JSON.stringify(group.specifications, null, 2) : ""
-  )
-
-  const onSubmit = async () => {
-    if (!name.trim()) {
-      toast.error("Name is required")
-      return
-    }
-    let specifications: Record<string, unknown> | undefined
-    if (specs.trim()) {
-      try {
-        specifications = JSON.parse(specs)
-      } catch {
-        toast.error("Specifications must be valid JSON")
-        return
-      }
-    }
-
-    const body: Record<string, unknown> = {
-      name: name.trim(),
-      status,
-      composition: composition.trim() || undefined,
-      unit_of_measure: unit,
-      stock_location_id: locationId || undefined,
-      unit_cost: unitCost !== "" ? Number(unitCost) : undefined,
-      cost_currency: currency.trim() || undefined,
-      lead_time_days: leadTime !== "" ? Number(leadTime) : undefined,
-      minimum_order_quantity: moq !== "" ? Number(moq) : undefined,
-      ...(specifications ? { specifications } : {}),
-    }
-
-    try {
-      await mutateAsync(body)
-      toast.success("Group updated")
-      setOpen(false)
-    } catch (e: any) {
-      toast.error(e?.message || "Failed to update group")
-    }
-  }
-
-  return (
-    <FocusModal open={open} onOpenChange={setOpen}>
-      <FocusModal.Trigger asChild>
-        <Button size="small" variant="secondary" type="button">
-          <PencilSquare className="text-ui-fg-subtle" />
-          Edit
-        </Button>
-      </FocusModal.Trigger>
-      <FocusModal.Content>
-        <FocusModal.Header>
-          <FocusModal.Title asChild>
-            <Heading>Edit group globals</Heading>
-          </FocusModal.Title>
-          <Button size="small" onClick={onSubmit} isLoading={isPending} type="button">
-            Save
-          </Button>
-        </FocusModal.Header>
-        <FocusModal.Body className="flex flex-col items-center overflow-y-auto py-8">
-          <div className="flex w-full max-w-2xl flex-col gap-y-4">
-            <Text size="small" className="text-ui-fg-subtle">
-              These specs are set once for the group. New colors inherit any field
-              they don't already have (fill-blank). To push changes onto colors
-              that already exist, run Settings → Data Plumbing →
-              &ldquo;Backfill group globals to colors&rdquo;.
-            </Text>
-
-            <div className="flex flex-col gap-y-1">
-              <Label size="small">Name</Label>
-              <Input value={name} onChange={(e) => setName(e.target.value)} />
-            </div>
-
-            <div className="grid grid-cols-2 gap-4">
-              <div className="flex flex-col gap-y-1">
-                <Label size="small">Status</Label>
-                <Select value={status} onValueChange={setStatus}>
-                  <Select.Trigger>
-                    <Select.Value />
-                  </Select.Trigger>
-                  <Select.Content>
-                    {STATUS_OPTIONS.map((s) => (
-                      <Select.Item key={s} value={s}>
-                        {s.replace("_", " ")}
-                      </Select.Item>
-                    ))}
-                  </Select.Content>
-                </Select>
-              </div>
-              <div className="flex flex-col gap-y-1">
-                <Label size="small">Unit of measure</Label>
-                <Select value={unit} onValueChange={setUnit}>
-                  <Select.Trigger>
-                    <Select.Value />
-                  </Select.Trigger>
-                  <Select.Content>
-                    {UNIT_OPTIONS.map((u) => (
-                      <Select.Item key={u} value={u}>
-                        {u}
-                      </Select.Item>
-                    ))}
-                  </Select.Content>
-                </Select>
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-y-1">
-              <Label size="small">Composition</Label>
-              <Input
-                value={composition}
-                onChange={(e) => setComposition(e.target.value)}
-                placeholder="e.g. 100% Cotton"
-              />
-            </div>
-
-            <div className="flex flex-col gap-y-1">
-              <Label size="small">Default receiving location</Label>
-              <Select value={locationId} onValueChange={setLocationId}>
-                <Select.Trigger>
-                  <Select.Value placeholder="Select a location" />
-                </Select.Trigger>
-                <Select.Content>
-                  {locations.map((sl: any) => (
-                    <Select.Item key={sl.id} value={sl.id}>
-                      {sl.name}
-                    </Select.Item>
-                  ))}
-                </Select.Content>
-              </Select>
-            </div>
-
-            <div className="grid grid-cols-2 gap-4">
-              <div className="flex flex-col gap-y-1">
-                <Label size="small">Default unit cost</Label>
-                <Input
-                  type="number"
-                  value={unitCost}
-                  onChange={(e) => setUnitCost(e.target.value)}
-                />
-              </div>
-              <div className="flex flex-col gap-y-1">
-                <Label size="small">Currency</Label>
-                <Input
-                  value={currency}
-                  onChange={(e) => setCurrency(e.target.value)}
-                  placeholder="e.g. inr"
-                />
-              </div>
-            </div>
-
-            <div className="grid grid-cols-2 gap-4">
-              <div className="flex flex-col gap-y-1">
-                <Label size="small">Lead time (days)</Label>
-                <Input
-                  type="number"
-                  value={leadTime}
-                  onChange={(e) => setLeadTime(e.target.value)}
-                />
-              </div>
-              <div className="flex flex-col gap-y-1">
-                <Label size="small">Minimum order quantity</Label>
-                <Input
-                  type="number"
-                  value={moq}
-                  onChange={(e) => setMoq(e.target.value)}
-                />
-              </div>
-            </div>
-
-            <div className="flex flex-col gap-y-1">
-              <Label size="small">Specifications (JSON)</Label>
-              <Textarea
-                value={specs}
-                onChange={(e) => setSpecs(e.target.value)}
-                rows={5}
-                placeholder='{ "gsm": 120, "weave": "poplin" }'
-              />
-            </div>
-          </div>
-        </FocusModal.Body>
-      </FocusModal.Content>
-    </FocusModal>
   )
 }
